requirements.txt
docker==4.4.4
j2cli==0.3.10
python-dateutil==2.8.2
hvac==0.7.2
Flask==1.1.4

requests~=2.27.1
Jinja2~=2.11.3
